rename package, remove unused imports
