+++ src/main/java/quiz/ColorPickerStreamHandler.java
@@ -11,20 +11,21 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package main.java.colorpicker;
+package main.java.quiz;
 
 import com.amazon.ask.Skill;
 import com.amazon.ask.SkillStreamHandler;
 import com.amazon.ask.Skills;
-import main.java.colorpicker.handlers.FallbackIntentHandler;
-import main.java.colorpicker.handlers.HelpIntentHandler;
-import main.java.colorpicker.handlers.LaunchRequestHandler;
-import main.java.colorpicker.handlers.SessionEndedRequestHandler;
-import main.java.colorpicker.handlers.WhatsMyColorIntentHandler;
-import main.java.colorpicker.handlers.CancelandStopIntentHandler;
-import main.java.colorpicker.handlers.MyColorIsIntentHandler;
-import main.java.colorpicker.handlers.StartQuizIntentHandler;
-import main.java.colorpicker.handlers.DefinePlayersIntentHandler;
+
+import main.java.quiz.handlers.CancelandStopIntentHandler;
+import main.java.quiz.handlers.DefinePlayersIntentHandler;
+import main.java.quiz.handlers.FallbackIntentHandler;
+import main.java.quiz.handlers.HelpIntentHandler;
+import main.java.quiz.handlers.LaunchRequestHandler;
+import main.java.quiz.handlers.MyColorIsIntentHandler;
+import main.java.quiz.handlers.SessionEndedRequestHandler;
+import main.java.quiz.handlers.StartQuizIntentHandler;
+import main.java.quiz.handlers.WhatsMyColorIntentHandler;
 
 public class ColorPickerStreamHandler extends SkillStreamHandler {
 	public static int num_players;
