+++ src/main/java/quiz/handlers/FallbackIntentHandler.java
@@ -1,4 +1,4 @@
-package main.java.colorpicker.handlers;
+package main.java.quiz.handlers;
 
 import com.amazon.ask.dispatcher.request.handler.HandlerInput;
 import com.amazon.ask.dispatcher.request.handler.RequestHandler;
